--- src/pages/Login.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <link rel="stylesheet" href="estilos.css">
-  </head>
-  <body>
-    <div class="container" id="container">
-      <div class="panel panel-izquierdo">
-        <h2 id="titulo">Crea tu cuenta</h2>
-        <p id="texto panel">
-          Registrate para ver mas informacion de nuestros eventos
-        </p>
-        <button id="registro">Registro</button>
-      </div>
-      <div class="panel panel-derecho">
-        <form class="form login">
-          <h2>LOGIN</h2>
-          <input type="email" placeholder="Email" required />
-          <input type="password" placeholder="password" required />
-          <button type="submit">Iniciar Seccion</button>
-        </form>
-        <form class="form registro">
-            <h2>REGISTRO</h2>
-            <input type="text" placeholder="Nombre" required>
-             <input type="email" placeholder="Email" required />
-             <input type="password" placeholder="password" required />
-             <button type="submit">Registro</button>
-        </form>
-      </div>
-
-    </div>
-    <script src="script.js"></script>
-  </body>
-</html>
