+++ src/pages/LoginRegistro.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Login from '../components/forms/Login/Login'
+
+export default function LoginRegistro() {
+  return (
+    <div className='container pt-5'>
+        <Login />
+    </div>
+  )
+}
